Lum-168 (Jira). Moved the 'sendOrderConfirmation' method to the mail service (removed the duplicate of mail service (mailOrder.ts file)) and replaced the default create order method (from the generic crud service) with an actual order creation method (the 'orders API' scope now contains only 1 create method instead of 2). Also changed the return type of almost all (except products) endpoints that can be executed by unauthorised or authorised users (changed related files in services, controllers and routes folders) to dto format (to avoid showing unnecessary or private information to users) and moved all checks from controllers to services.
