+++ src/api/services/auxiliary/mailService.ts
@@ -7,6 +7,18 @@ export const mailService = {
         const mailOptions = getMailOptions(to, subject, message);
         transporter.sendMail(mailOptions);
     },
+
+    async sendOrderConfirmation(email: string, orderDetails: any) {
+        const mailOptions = {
+            from: process.env.SMTP_EMAIL,
+            to: email,
+            subject: "Your Order Confirmation",
+            text: `Thank you for your order! Here are the details:\n\n${JSON.stringify(orderDetails, null, 2)}`,
+        };
+
+        const transporter = getTransporter();
+        await transporter.sendMail(mailOptions);
+    }
 }
 
 function getMailOptions(to: String, subject: String, message: String): Object {
