+++ src/api/services/cartService.ts
@@ -1,14 +1,21 @@
-import mongoose, { HydratedDocument } from "mongoose";
-import Cart, { ICart } from "../models/carts";
+import mongoose from "mongoose";
+import Cart from "../models/carts";
 import ApiError from "../errors/ApiError";
 import { StatusCodes } from "http-status-codes";
 import NotFoundError from "../errors/general/NotFoundError";
 import Product from "../models/products";
 import { jwtService } from "./auxiliary/jwtService";
 import { ensureItemExists } from "./genericCrudService";
+import ValidationError from "../errors/validation/ValidationError";
+import { CartDTO, convertToCartDTO } from "../dto/CartDTO";
 
 export const cartService = {
-    addProduct: async (bearerToken: string | undefined, productId: mongoose.Types.ObjectId, quantity: number): Promise<HydratedDocument<ICart>> => {
+    addProduct: async (bearerToken: string | undefined, productId: mongoose.Types.ObjectId, quantity: number): Promise<CartDTO> => {
+
+        if (!productId || !quantity) {
+            throw new ValidationError('ProductId, and quantity are required');
+        }
+
         const user = await jwtService.getUserFromBearerToken(bearerToken);
         await ensureItemExists(Product, '_id', productId);
 
@@ -34,10 +41,20 @@ export const cartService = {
         }
 
         // Зберегти кошик
-        return await cart.save();
+        const savedCart = await cart.save();
+        return convertToCartDTO(savedCart);
     },
 
-    removeProduct: async (bearerToken: string | undefined, productId: string, quantity: number): Promise<HydratedDocument<ICart>> => {
+    removeProduct: async (bearerToken: string | undefined, productId: string, quantity: number): Promise<CartDTO> => {
+
+        if (!productId) {
+            throw new ValidationError('ProductId is required');
+        }
+
+        if (quantity < 0) {
+            throw new ValidationError('Quantity cannot be negative');
+        }
+
         // Встановити значення за замовчуванням для quantity, якщо воно не вказане
         const removeQuantity = quantity || 1;
         const user = await jwtService.getUserFromBearerToken(bearerToken);
@@ -63,7 +80,7 @@ export const cartService = {
         }
 
         // Зберегти кошик
-        return await cart.save();
-
+        const savedCart = await cart.save();
+        return convertToCartDTO(savedCart);
     }
 }
