+++ src/config/swagger/swaggerOptions.ts
@@ -108,10 +108,14 @@ const options = {
                         properties: {
                             totalPrice: { type: 'number', example: 1000 },
                             products: {
-                                type: 'object', properties: [{
-                                    productId: { type: 'string', example: '67b1cf2abb50ec824a8390de' },
-                                    quantity: { type: 'number', example: 1 }
-                                }]
+                                type: 'array',
+                                items: {
+                                    type: 'object',
+                                    properties: {
+                                        productId: { type: 'string', example: '67b1cf2abb50ec824a8390de' },
+                                        quantity: { type: 'number', example: 1 }
+                                    }
+                                }
                             }
                         }
                     },
